[#19] rename: 폴더명, 파일명, mixin명 변경
