--- components/login/kakaoButton/kakaoButton.module.scss
@@ -1,6 +1,0 @@
-@import '/components/common/Button/Button.module.scss';
-
-.kakaoButton {
-  @include RoundButton(var(--kakao-background), var(--kakao-color));
-  gap: 14px;
-}
