+++ pages/login/index.tsx
@@ -1,6 +1,6 @@
 import styles from './index.module.scss';
-import LoginTop from '@/components/login/loginTop/loginTop';
-import LoginButtons from '@/components/login/loginButtons/loginButtons';
+import LoginTop from '@/components/login/login-top/login-top';
+import LoginButtons from '@/components/login/login-buttons/login-buttons';
 
 export default function Login() {
   return (
